refactor: centralize full paper generation
